openenv_lifeops/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-lifeops-env"
version = "0.1.0"
description = "LifeOps schedule management environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core==0.2.1",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
]

[tool.setuptools.packages.find]
where = [".."]
include = ["openenv_lifeops*", "env*"]
